@@ -12,6 +12,8 @@ var _get = function get(_x, _x2, _x3) { var _again = true; _function: while (_ag
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { 'default': obj }; }
 
+function _objectWithoutProperties(obj, keys) { var target = {}; for (var i in obj) { if (keys.indexOf(i) >= 0) continue; if (!Object.prototype.hasOwnProperty.call(obj, i)) continue; target[i] = obj[i]; } return target; }
+
 function _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError('Cannot call a class as a function'); } }
 
 function _inherits(subClass, superClass) { if (typeof superClass !== 'function' && superClass !== null) { throw new TypeError('Super expression must either be null or a function, not ' + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }
@@ -24,6 +26,10 @@ var _libFilter_props_from = require('./lib/filter_props_from');
 
 var _libExtract_data_from = require('./lib/extract_data_from');
 
+var _libDetermine_row_span = require('./lib/determine_row_span');
+
+var _libExtend = require('./lib/extend');
+
 var _unsafe = require('./unsafe');
 
 var _thead = require('./thead');
@@ -35,6 +41,10 @@ var _tr = require('./tr');
 var _tfoot = require('./tfoot');
 
 var _paginator = require('./paginator');
+
+var _propTypes = require("prop-types");
+
+var _propTypes2 = _interopRequireDefault(_propTypes);
 
 var Table = (function (_React$Component) {
     _inherits(Table, _React$Component);
@@ -58,6 +68,10 @@ var Table = (function (_React$Component) {
             var sortingColumn = props.sortBy || props.defaultSort;
             this.state.currentSort = this.getCurrentSort(sortingColumn);
         }
+
+        // used for generating a key for a <Tr> that will change <Td> counts -- avoids jitters in the update
+        this.cloneIndex = 0;
+        this.renderNoDataComponent = this.renderNoDataComponent.bind(this);
     }
 
     _createClass(Table, [{
@@ -125,12 +139,17 @@ var Table = (function (_React$Component) {
                                     } else if (typeof descendant.props.children !== 'undefined') {
                                         value = descendant.props.children;
                                     } else {
-                                        console.warn('exports.Td specified without ' + 'a `data` property or children, ' + 'ignoring');
+                                        var warning = 'exports.Td specified without ' + 'a `data` property or children, ignoring';
+                                        if (typeof descendant.props.column !== 'undefined') {
+                                            warning += '. See definition for column \'' + descendant.props.column + '\'.';
+                                        }
+                                        console.warn(warning);
                                         return;
                                     }
 
                                     childData[descendant.props.column] = {
                                         value: value,
+                                        component: descendant.type,
                                         props: (0, _libFilter_props_from.filterPropsFrom)(descendant.props),
                                         __reactableMeta: true
                                     };
@@ -304,37 +323,174 @@ var Table = (function (_React$Component) {
     }, {
         key: 'applyFilter',
         value: function applyFilter(filter, children) {
+            var _this = this;
+
             // Helper function to apply filter text to a list of table rows
             filter = filter.toLowerCase();
-            var matchedChildren = [];
 
-            for (var i = 0; i < children.length; i++) {
-                var data = children[i].props.data;
+            /**
+             * We face two problems when trying to filter with rowSpan:
+             *  1) what happens when the row that specifies the <Td> with the row span is filtered out?
+             *       move the rowSpan definition to the next visible row and adjust the rowSpan value to exclude
+             *                 all the filtered out rows (e.g. if we have a rowSpan=5 and the 1st and 5th row are filtered
+             *                 out, the rowSpan property & value shifts to row 2 and rowSpan now becomes 3)
+             *
+             *  2) what happens when the row/cell that specifies the rowSpan matches the text:
+             *       immediately include all the subsequent rows the that the <Td> spans in the result list.
+             *
+             */
 
-                for (var filterColumn in this._filterable) {
+            var matchedChildren = {}; // row index => the matched data
+
+            // houses the rowSpan definition for each row/columnName combo (e.g. [0]['Name'] provides the rowSpan definition
+            //  for the item at row 0 in the 'Name' column
+            var rowSpanReferences = {};
+
+            // the current rowSpan definition for each column, columnName => rowSpan definition
+            var currentRowSpanState = {};
+
+            var _loop = function (i) {
+                var child = children[i];
+                var data = child.props.data;
+
+                // look at each column in this row and see if there are rowSpan properties
+                Object.keys(data).forEach(function (k) {
+                    if (typeof k !== 'undefined' && typeof data[k] !== 'undefined') {
+                        var state = currentRowSpanState[k];
+
+                        // check if we've exhausted our specified number of rows for this column
+                        if (typeof state !== 'undefined') {
+                            var rowSpanEnd = state.startsAt + state.rowSpan - 1;
+                            if (i >= rowSpanEnd) {
+                                delete currentRowSpanState[k];
+                            }
+                        }
+
+                        var rowSpan = (0, _libDetermine_row_span.determineRowSpan)(child, k);
+                        // we don't want to waste our time with single rows, only keep a state if we have >1
+                        if (rowSpan > 1) {
+                            state = {
+                                rowSpan: rowSpan,
+                                startsAt: i,
+                                component: child,
+                                columnName: k
+                            };
+                            currentRowSpanState[k] = state;
+
+                            // store the reference to the rowSpanDefinition for all columns it's relevant to
+                            var rowSpanEnd = i + rowSpan;
+                            for (var idx = i; idx < rowSpanEnd && idx < children.length; idx++) {
+                                if (typeof rowSpanReferences[idx] === 'undefined') {
+                                    rowSpanReferences[idx] = {};
+                                }
+                                rowSpanReferences[idx][k] = state;
+                            }
+                        }
+                    }
+                });
+
+                // 2) look through all 'filterable' columns for this row and see if we can find the filterBy text
+                for (var filterColumn in _this._filterable) {
                     if (typeof data[filterColumn] !== 'undefined') {
                         // Default filter
-                        if (typeof this._filterable[filterColumn] === 'undefined' || this._filterable[filterColumn] === 'default') {
+                        if (typeof _this._filterable[filterColumn] === 'undefined' || _this._filterable[filterColumn] === 'default') {
+
                             if ((0, _libExtract_data_from.extractDataFrom)(data, filterColumn).toString().toLowerCase().indexOf(filter) > -1) {
-                                matchedChildren.push(children[i]);
+
+                                // upon a filter text match, we include the cell that matched and all the rows that
+                                //  the cell spans
+                                var rowSpan = (0, _libDetermine_row_span.determineRowSpan)(child, filterColumn);
+                                var rowSpanEnd = i + rowSpan;
+
+                                for (idx = i; idx < rowSpanEnd && idx < children.length; idx++) {
+                                    otherChild = children[idx];
+
+                                    matchedChildren[idx] = otherChild;
+
+                                    // we no longer need to handle rowSpan for the other rows
+                                    if (typeof rowSpanReferences[idx] !== 'undefined') {
+                                        delete rowSpanReferences[idx][filterColumn];
+                                    }
+                                }
+
+                                // we've rendered all the rows this cell spans, we no longer need to maintain state
+                                delete currentRowSpanState[filterColumn];
+
                                 break;
                             }
                         } else {
                             // Apply custom filter
-                            if (this._filterable[filterColumn]((0, _libExtract_data_from.extractDataFrom)(data, filterColumn).toString(), filter)) {
-                                matchedChildren.push(children[i]);
+                            if (_this._filterable[filterColumn]((0, _libExtract_data_from.extractDataFrom)(data, filterColumn).toString(), filter)) {
+                                matchedChildren[i] = child;
                                 break;
                             }
                         }
                     }
                 }
+            };
+
+            for (var i = 0; i < children.length; i++) {
+                var idx;
+                var otherChild;
+
+                _loop(i);
             }
 
-            return matchedChildren;
+            // Now that we know which results will be displayed, we can calculate a rowSpan that will encompass
+            //  only the visible rows.
+            // Iterate through the matched children and check if there are rowSpan modifications they need to make. Hint: we
+            //  stored all row/col combos needing a modification in rowSpanReferences
+            var resultingChildren = Object.keys(matchedChildren).map((function (rowIndex) {
+                var definitions = rowSpanReferences[rowIndex];
+                if (typeof definitions !== 'undefined') {
+
+                    // the child we will clone and give a new <Td>
+                    var child = matchedChildren[rowIndex];
+
+                    // we'll be recreating this <Tr> with a rowSpan property carried over from the <Tr> with the rowSpan
+                    var newData = (0, _libExtend.extend)({}, child.props.data);
+                    var newProps = (0, _libExtend.extend)({}, child.props);
+                    newProps.data = newData;
+
+                    Object.keys(definitions).forEach(function (columnName) {
+                        var definition = definitions[columnName];
+                        // the tr with the rowSpan property
+                        var tr = definition.component;
+
+                        // this child was a part of a rowSpan
+                        var rowSpanEnd = definition.startsAt + definition.rowSpan;
+                        var remainingRowSpan = rowSpanEnd - rowIndex;
+
+                        // if the rows that we're supposed to span did not match the filter text, we need to reduce the remainingRowSpan
+                        for (var idx = rowIndex; idx < rowSpanEnd; idx++) {
+                            if (typeof matchedChildren[idx] === 'undefined') {
+                                remainingRowSpan--;
+                            }
+                        }
+
+                        // clone the data for the column to avoid altering the original props (e.g. don't want to change the rowSpan on original Td)
+                        newProps.data[columnName] = (0, _libExtend.extend)({}, tr.props.data[columnName]);
+                        newProps.data[columnName].props = (0, _libExtend.extend)({}, newProps.data[columnName].props);
+                        newProps.data[columnName].props.rowSpan = remainingRowSpan;
+                    });
+
+                    // new key to avoid jitters, we want a brand new component
+                    newProps.key = "cloned-" + _this.cloneIndex++;
+                    newProps.renderState = child.props.renderState;
+
+                    return _react2['default'].cloneElement(child, newProps);
+                } else {
+                    return matchedChildren[rowIndex];
+                }
+            }).bind(this));
+
+            return resultingChildren;
         }
     }, {
         key: 'sortByCurrentSort',
         value: function sortByCurrentSort() {
+            var _this2 = this;
+
             // Apply a sort function according to the current sort in the state.
             // This allows us to perform a default sort even on a non sortable column.
             var currentSort = this.state.currentSort;
@@ -343,34 +499,129 @@ var Table = (function (_React$Component) {
                 return;
             }
 
-            this.data.sort((function (a, b) {
-                var keyA = (0, _libExtract_data_from.extractDataFrom)(a, currentSort.column);
-                keyA = (0, _unsafe.isUnsafe)(keyA) ? keyA.toString() : keyA || '';
-                var keyB = (0, _libExtract_data_from.extractDataFrom)(b, currentSort.column);
-                keyB = (0, _unsafe.isUnsafe)(keyB) ? keyB.toString() : keyB || '';
+            // we'll be splitting each set of data into buckets, then sorting them. A 'set' is defined as rows sharing a
+            //  <Td> with a rowSpan
+            var singles = []; // rows without any rowSpans
+            var buckets = {};
+            var bucketIndex = 0;
 
-                // Default sort
-                if (typeof this._sortable[currentSort.column] === 'undefined' || this._sortable[currentSort.column] === 'default') {
+            // helps us keep track of where we are in the last rowSpan we saw
+            var currentRowSpan = null;
 
-                    // Reverse direction if we're doing a reverse sort
-                    if (keyA < keyB) {
-                        return -1 * currentSort.direction;
-                    }
+            for (var idx = 0; idx < this.data.length; idx++) {
+                var obj = this.data[idx];
 
-                    if (keyA > keyB) {
-                        return 1 * currentSort.direction;
-                    }
-
-                    return 0;
-                } else {
-                    // Reverse columns if we're doing a reverse sort
-                    if (currentSort.direction === 1) {
-                        return this._sortable[currentSort.column](keyA, keyB);
-                    } else {
-                        return this._sortable[currentSort.column](keyB, keyA);
+                // check if rowSpan is completed
+                if (currentRowSpan != null) {
+                    var rowSpanEnd = currentRowSpan.startsAt + currentRowSpan.rowSpan;
+                    if (rowSpanEnd <= idx) {
+                        bucketIndex++;
+                        currentRowSpan = null;
                     }
                 }
-            }).bind(this));
+
+                // make sure the bucket is defined
+                if (typeof buckets[bucketIndex] === 'undefined') {
+                    buckets[bucketIndex] = [];
+                }
+
+                if (typeof obj.data !== 'undefined') {
+                    // find the column with the next biggest rowSpan
+                    var rowSpanOfSortColumn = (0, _libDetermine_row_span.determineRowSpan)(obj.data, currentSort.column);
+
+                    if (rowSpanOfSortColumn !== 1) {
+                        // not supported
+                        console.warn("Cannot sort by column '" + currentSort.column + "', sorting by columns that have cells with rowSpan is currently not supported. See https://github.com/glittershark/reactable/issues/84");
+                        return;
+                    }
+
+                    for (var col in obj.data) {
+
+                        if (col === currentSort.column) {
+                            // ignore the sort column
+                            continue;
+                        }
+
+                        var rowSpanCurrentCol = (0, _libDetermine_row_span.determineRowSpan)(obj.data, col);
+
+                        // if the rowSpan we found is less than the current debt but greater than the rowSpan of our column,
+                        //      we will use that number
+                        if (rowSpanCurrentCol > rowSpanOfSortColumn && (currentRowSpan == null || currentRowSpan.rowSpan > rowSpanCurrentCol)) {
+                            currentRowSpan = {
+                                startsAt: idx,
+                                rowSpan: rowSpanCurrentCol
+                            };
+                        }
+                    }
+                }
+
+                if (currentRowSpan == null) {
+                    singles.push(obj);
+                } else {
+                    buckets[bucketIndex].push(obj);
+                }
+            }
+
+            // run a sort on each bucket
+            buckets = Object.keys(buckets).map(function (dataSet) {
+                return buckets[dataSet].sort(_this2.sortFunction.bind(_this2));
+            });
+            buckets.push(singles.sort(this.sortFunction.bind(this)));
+
+            // flatten
+            this.data = buckets.reduce(function (a, current) {
+                // look for row spans and put the definition of the <Td> on the top row after sort
+                if (current.length > 0) {
+                    for (var idx = 0; idx < current.length; idx++) {
+                        var obj = current[idx];
+
+                        for (var col in obj.data) {
+                            var rowSpan = (0, _libDetermine_row_span.determineRowSpan)(obj.data, col);
+                            var currentPosition = idx % rowSpan;
+                            var newRowSpanIdx = idx - currentPosition;
+
+                            if (idx !== newRowSpanIdx) {
+                                // need to push the rowSpan property & values to the top row
+                                current[newRowSpanIdx].data[col] = (0, _libExtend.extend)({}, obj.data[col]);
+                                delete obj.data[col];
+                            }
+                        }
+                    }
+                }
+                return a.concat(current);
+            }, []);
+        }
+    }, {
+        key: 'sortFunction',
+        value: function sortFunction(a, b) {
+            var currentSort = this.state.currentSort;
+
+            var keyA = (0, _libExtract_data_from.extractDataFrom)(a, currentSort.column);
+            keyA = (0, _unsafe.isUnsafe)(keyA) ? keyA.toString() : keyA || '';
+            var keyB = (0, _libExtract_data_from.extractDataFrom)(b, currentSort.column);
+            keyB = (0, _unsafe.isUnsafe)(keyB) ? keyB.toString() : keyB || '';
+
+            // Default sort
+            if (typeof this._sortable[currentSort.column] === 'undefined' || this._sortable[currentSort.column] === 'default') {
+
+                // Reverse direction if we're doing a reverse sort
+                if (keyA < keyB) {
+                    return -1 * currentSort.direction;
+                }
+
+                if (keyA > keyB) {
+                    return 1 * currentSort.direction;
+                }
+
+                return 0;
+            } else {
+                // Reverse columns if we're doing a reverse sort
+                if (currentSort.direction === 1) {
+                    return this._sortable[currentSort.column](keyA, keyB);
+                } else {
+                    return this._sortable[currentSort.column](keyB, keyA);
+                }
+            }
         }
     }, {
         key: 'onSort',
@@ -398,9 +649,29 @@ var Table = (function (_React$Component) {
             }
         }
     }, {
+        key: 'renderNoDataComponent',
+        value: function renderNoDataComponent(columns) {
+            var noDataFunc = this.props.noDataComponent;
+            if (typeof noDataFunc === 'function') {
+                return noDataFunc(columns);
+            } else if (this.props.noDataText) {
+                return _react2['default'].createElement(
+                    'tr',
+                    { className: 'reactable-no-data' },
+                    _react2['default'].createElement(
+                        'td',
+                        { colSpan: columns.length },
+                        this.props.noDataText
+                    )
+                );
+            } else {
+                return null;
+            }
+        }
+    }, {
         key: 'render',
         value: function render() {
-            var _this = this;
+            var _this3 = this;
 
             var children = [];
             var columns = undefined;
@@ -430,6 +701,7 @@ var Table = (function (_React$Component) {
 
             // Build up table rows
             if (this.data && typeof this.data.map === 'function') {
+                var renderState = {};
                 // Build up the columns array
                 children = children.concat(this.data.map((function (rawData, i) {
                     var data = rawData;
@@ -462,7 +734,7 @@ var Table = (function (_React$Component) {
                         }
                     }
 
-                    return _react2['default'].createElement(_tr.Tr, _extends({ columns: columns, key: i, data: data }, props));
+                    return _react2['default'].createElement(_tr.Tr, _extends({ columns: columns, key: i, data: data }, props, { renderState: renderState }));
                 }).bind(this)));
             }
 
@@ -505,26 +777,21 @@ var Table = (function (_React$Component) {
             }
 
             // Manually transfer props
-            var props = (0, _libFilter_props_from.filterPropsFrom)(this.props);
 
-            var noDataText = this.props.noDataText ? _react2['default'].createElement(
-                'tr',
-                { className: 'reactable-no-data' },
-                _react2['default'].createElement(
-                    'td',
-                    { colSpan: columns.length },
-                    this.props.noDataText
-                )
-            ) : null;
+            var _filterPropsFrom = (0, _libFilter_props_from.filterPropsFrom)(this.props);
+
+            var noDataComponent = _filterPropsFrom.noDataComponent;
+
+            var props = _objectWithoutProperties(_filterPropsFrom, ['noDataComponent']);
 
             var tableHeader = null;
             if (columns && columns.length > 0 && showHeaders) {
                 tableHeader = _react2['default'].createElement(_thead.Thead, { columns: columns,
                     filtering: filtering,
                     onFilter: function (filter) {
-                        _this.setState({ filter: filter });
-                        if (_this.props.onFilter) {
-                            _this.props.onFilter(filter);
+                        _this3.setState({ filter: filter });
+                        if (_this3.props.onFilter) {
+                            _this3.props.onFilter(filter);
                         }
                     },
                     filterPlaceholder: this.props.filterPlaceholder,
@@ -542,16 +809,16 @@ var Table = (function (_React$Component) {
                 _react2['default'].createElement(
                     'tbody',
                     { className: 'reactable-data', key: 'tbody' },
-                    currentChildren.length > 0 ? currentChildren : noDataText
+                    currentChildren.length > 0 ? currentChildren : this.renderNoDataComponent(columns)
                 ),
                 pagination === true ? _react2['default'].createElement(_paginator.Paginator, { colSpan: columns.length,
                     pageButtonLimit: pageButtonLimit,
                     numPages: numPages,
                     currentPage: currentPage,
                     onPageChange: function (page) {
-                        _this.setState({ currentPage: page });
-                        if (_this.props.onPageChange) {
-                            _this.props.onPageChange(page);
+                        _this3.setState({ currentPage: page });
+                        if (_this3.props.onPageChange) {
+                            _this3.props.onPageChange(page);
                         }
                     },
                     previousPageLabel: this.props.previousPageLabel,
@@ -574,4 +841,15 @@ Table.defaultProps = {
     itemsPerPage: 0,
     filterBy: '',
     hideFilterInput: false
+};
+
+Table.propTypes = {
+    sortBy: _propTypes2['default'].bool,
+    itemsPerPage: _propTypes2['default'].number, // number of items to display per page
+    filterable: _propTypes2['default'].array, // columns to look at when applying the filter specified by filterBy
+    filterBy: _propTypes2['default'].string, // text to filter the results by (see filterable)
+    hideFilterInput: _propTypes2['default'].bool, // Whether the default input field for the search/filter should be hidden or not
+    hideTableHeader: _propTypes2['default'].bool, // Whether the table header should be hidden or not
+    noDataText: _propTypes2['default'].string, // Text to be displayed in the event there is no data to show
+    noDataComponent: _propTypes2['default'].func // function called to provide a component to display in the event there is no data to show (supercedes noDataText)
 };
