@@ -89,7 +89,7 @@ var Thead = (function (_React$Component) {
                         onKeyDown: this.handleKeyDownTh.bind(this, column),
                         role: thRole,
                         tabIndex: '0' }),
-                    column.label
+                    column.content || column.label
                 ));
             }
 
@@ -126,7 +126,8 @@ var Thead = (function (_React$Component) {
 
                     // use the content as the label & key
                     if (typeof th.props.children !== 'undefined') {
-                        column.label = th.props.children;
+                        column.label = th.props.label || th.props.children;
+                        column.content = th.props.children || th.props.label;
                         column.key = column.label;
                     }
 
